+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## To Be Released
 
 * feat: add support to list container sizes [#234](https://github.com/Scalingo/go-scalingo/pull/234)
+* feat(apps): add AppsRouterLogs function [#236](https://github.com/Scalingo/go-scalingo/pull/236)
 
 ## 4.14.3